refactor: 인기 테마 조회 api /reservations/popular-themes에서 themes/popular로 변경
