+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,7 +1,6 @@
 package roomescape.service;
 
 import org.springframework.stereotype.Service;
-import roomescape.controller.reservation.PopularThemeResponse;
 import roomescape.controller.reservation.ReservationRequest;
 import roomescape.controller.reservation.ReservationResponse;
 import roomescape.domain.Reservation;
@@ -15,7 +14,6 @@ import roomescape.service.exception.PreviousTimeException;
 import roomescape.service.exception.ThemeNotFoundException;
 import roomescape.service.exception.TimeNotFoundException;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -81,12 +79,5 @@ public class ReservationService {
 
     public int deleteReservation(final Long id) {
         return reservationRepository.deleteById(id);
-    }
-
-    public List<PopularThemeResponse> getPopularThemes(final LocalDate today) {
-        final List<Theme> reservations = reservationRepository.findPopularThemes(today);
-        return reservations.stream()
-                .map(PopularThemeResponse::from)
-                .toList();
     }
 }
