+++ src/main/java/roomescape/service/ThemeService.java
@@ -1,13 +1,15 @@
 package roomescape.service;
 
 import org.springframework.stereotype.Service;
+import roomescape.controller.reservation.PopularThemeResponse;
 import roomescape.controller.theme.CreateThemeRequest;
 import roomescape.controller.theme.CreateThemeResponse;
 import roomescape.domain.Theme;
-import roomescape.service.exception.ThemeUsedException;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ThemeRepository;
+import roomescape.service.exception.ThemeUsedException;
 
+import java.time.LocalDate;
 import java.util.List;
 
 @Service
@@ -38,5 +40,12 @@ public class ThemeService {
             throw new ThemeUsedException("예약된 테마는 삭제할 수 없습니다.");
         }
         return themeRepository.delete(id);
+    }
+
+    public List<PopularThemeResponse> getPopularThemes(final LocalDate today) {
+        final List<Theme> reservations = reservationRepository.findPopularThemes(today);
+        return reservations.stream()
+                .map(PopularThemeResponse::from)
+                .toList();
     }
 }
