+++ src/main/resources/static/js/ranking.js
@@ -1,8 +1,5 @@
 document.addEventListener('DOMContentLoaded', () => {
-  /*
-  TODO: [3단계] 인기 테마 - 인기 테마 목록 조회 API 호출
-  */
-  requestRead('/reservations/popular-themes') // 인기 테마 목록 조회 API endpoint
+  requestRead('/themes/popular') // 인기 테마 목록 조회 API endpoint
     .then(render)
     .catch(error => console.error('Error fetching times:', error));
 });
@@ -10,10 +7,6 @@ document.addEventListener('DOMContentLoaded', () => {
 function render(data) {
   const container = document.getElementById('theme-ranking');
 
-  /*
-  TODO: [3단계] 인기 테마 - 인기 테마 목록 조회 API 호출 후 렌더링
-        response 명세에 맞춰 name, thumbnail, description 값 설정
-  */
   data.forEach(theme => {
     const name = theme.name;
     const thumbnail = theme.thumbnail;
